backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.23.0
sqlmodel>=0.0.8
python-multipart>=0.0.6
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
psycopg[binary]>=3.3.3
psycopg2-binary>=2.9.9
python-dotenv>=1.1.1
pydantic-settings>=2.0.0
